avoid int overflow on multiplication

+++ src/DualVNH5019MotorShield.cpp
@@ -239,7 +239,7 @@ tools::mean_std_t DualVNH5019MotorShield::get_mean_std(int motor, int samples){
   tmp.mean = 0.;
   tmp.std = 0.;
 
-  unsigned int foo;
+  double foo;
   for (int i=0; i<samples; i++)
   {
     foo = getCurrentMilliamps(motor);
